feat: refactoring of the code to prevent processing bookmark that should not be (and re-organization of the add-on's structure)
BREAKING CHANGES: in addition to re-organizing the structure of the add-on, this commit also introduces a new way of handling and processing bookmarks. This is done in order to attempt to resolve the issues that have been introduced with version 3 of the addon and due to the internal changes introduced by Firefox at the time (as well as ongoing API limitations).

This commit also contains additional minor improvements throughout the codebase, mostly to enhance development and user experience.